Added CommentBox component and moved all comment related components to it for seperation of concern
